requirements.txt
attrs==19.3.0
Django==2.2.9
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-js-asset==1.2.2
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.8.0
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.1
